requirements.txt
Flask==3.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0

